:- completed integration ‘execLimitOrder’ (+ *** additional changes that i can’t remember ***) 	- works with limit sell and limit buy 	- added bin_clientGoLimitBuy
:- added flags 11, 12, and 13
	—-set-ls
	—-c
	—-term-launch

NOTE: there appears to be some code updates in this commit that i can’t remember what they are tied to. However, executing limit order sells and buys seem to be working fine. And executing market order buys and sells seem to be working fine as well.
NOTE_2: not exactly sure what term launch (—-term-launch flag) is for